refactor: JobInfoItemDetail -> jobDetail 으로 수정
