+++ front-end/src/domains/Job.ts
@@ -1,5 +1,5 @@
 export { default as JobInfoItem } from "./JobInfoItem/JobInfoItem";
-export { default as JobInfoItemDetail } from "./JobInfoItemDetail/JobInfoItemDetail";
+export { default as JobDetail } from "./JobDetail/JobDetail";
 export { default as JobInfoList } from "./JobInfoList/JobInfoList";
 export { default as JobRegisterForm } from "./JobRegisterForm/JobRegisterForm";
 export { default as JobRegisterRadioGroup } from "./JobRegisterRadioGroup/JobRegisterRadioGroup";
