+++ front-end/src/domains/JobDetail/useJobDetail.ts
@@ -8,17 +8,6 @@ export const useGoToPage = (pointer: number) => {
   return () => history.go(pointer);
 };
 
-export const useJobDetail = ({ labId, jobId }: { labId: number; jobId: number }) => {
-  const { status, makeRequest, data } = useGetJobDetail({ labId, jobId });
-
-  useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    makeRequest();
-  }, [makeRequest]);
-
-  return { status, detail: data };
-};
-
 export const useLabId = () => {
   const { myInfo } = useAuth();
 
@@ -29,4 +18,15 @@ export const useJobId = () => {
   const { jobId } = useParams<{ jobId?: string }>();
 
   return Number(jobId);
+};
+
+export const useJobDetail = ({ labId, jobId }: { labId: number; jobId: number }) => {
+  const { status, makeRequest, data } = useGetJobDetail({ labId, jobId });
+
+  useEffect(() => {
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    makeRequest();
+  }, [makeRequest]);
+
+  return { status, detail: data };
 };
