+++ front-end/src/pages/manager/JobViewDetail/JobViewDetail.tsx
@@ -1,13 +1,13 @@
 import { Layout } from "../../../components";
 import { ManagerNavigation, ManagerHeader, ManagerSubHeader } from "../../../domains/Manager";
-import { JobInfoItemDetail } from "../../../domains/Job";
+import { JobDetail } from "../../../domains/Job";
 
 const JobViewDetail = () => (
   <Layout
     Header={<ManagerHeader />}
     SubHeader={<ManagerSubHeader />}
     Navigation={<ManagerNavigation />}
-    Content={<JobInfoItemDetail />}
+    Content={<JobDetail />}
   />
 );
